+++ cypress/features/regression/designSystem/fieldset.feature
@@ -1,9 +1,9 @@
-Feature: Experimental Fieldset component
-  I want to check Experimental Fieldset component properties
+Feature: Design System Fieldset component
+  I want to check Design System Fieldset component properties
 
   @positive
   Scenario Outline: Change legend in Fieldset to <legend>
-    When I open default "Experimental Fieldset Test" component in noIFrame with "fieldset" json from "experimental" using "<nameOfObject>" object name
+    When I open default "Design System Fieldset Test" component in noIFrame with "fieldset" json from "designSystem" using "<nameOfObject>" object name
     Then legend on preview is <legend>
     Examples:
       | legend                       | nameOfObject           |
@@ -12,5 +12,5 @@ Feature: Experimental Fieldset component
 
   @negative
   Scenario: Set legend in Fieldset to empty
-    When I open default "Experimental Fieldset Test" component in noIFrame with "fieldset" json from "experimental" using "legendEmpty" object name
+    When I open default "Design System Fieldset Test" component in noIFrame with "fieldset" json from "designSystem" using "legendEmpty" object name
     Then legend on preview not exists
