+++ src/components/dialog/dialog.stories.mdx
@@ -11,7 +11,7 @@ import {
   RadioButton,
   RadioButtonGroup,
 } from "../../__experimental__/components/radio-button";
-import Fieldset from "../../__experimental__/components/fieldset";
+import Fieldset from "../fieldset";
 import Loader from "../loader";
 
 <Meta title="Dialog" parameters={{ info: { disable: true } }} />
