+++ src/components/form/form.spec.js
@@ -16,7 +16,7 @@ import FormSummary from "./form-summary.component";
 import { StyledFormSummary, StyledInternalSummary } from "./form-summary.style";
 import Icon from "../icon";
 import Button from "../button";
-import { FieldsetStyle } from "../../__experimental__/components/fieldset/fieldset.style";
+import { FieldsetStyle } from "../fieldset/fieldset.style";
 
 describe("Form", () => {
   let wrapper;
