+++ src/components/show-edit-pod/show-edit-pod.stories.js
@@ -6,7 +6,7 @@ import ShowEditPod, { BaseShowEditPod } from "./show-edit-pod.component";
 import Content from "../content";
 
 import Textbox from "../textbox";
-import Fieldset from "../../__experimental__/components/fieldset";
+import Fieldset from "../fieldset";
 
 export default {
   title: "ShowEditPod/Test",
